feat: Add level selection screen
This commit introduces the level selection screen, allowing players to choose between different levels before starting the game.

- Adds a new `LevelSelectionUIController` to manage the UI for the level selection screen.
- The screen includes buttons for "Level One", "Level Two", and a "Menu" button to return to the main menu.
- Integrates the new UI controller into the `UIService` to be displayed at the correct game state.
- Updates the main menu's "Play" button to navigate to the new level selection screen.

+++ Linked-List-Snake/source/Level/LevelModel.cpp
@@ -8,8 +8,8 @@ namespace Level
 
 	void LevelModel::initialize(int width, int height)
 	{
-		cell_width = width / number_of_columns;
-		cell_height = height / number_of_rows;
+		cell_width = static_cast<float>(width) / number_of_columns;
+		cell_height = static_cast<float>(height) / number_of_rows;
 	}
 
 	float LevelModel::getCellWidth()
